+++ tree/suite_test.go
@@ -1,0 +1,13 @@
+package tree
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestTreeIntegration(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "tree integration suite")
+}